requirements.txt
bson==0.5.10
Django==4.0.3
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
requests==2.27.1